third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-valid-relative-color-expected.txt

This is a testharness.js-based test.
Found 4 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] e.style['color'] = "rgb(from rebeccapurple r calc(g * .5 + g * .5) 10)" should set the property value
  Colors do not match.\nActual:   rgb(from rebeccapurple r calc(0.5 * g + 0.5 * g) 10)\nExpected: rgb(from rebeccapurple r calc((0.5 * g) + (0.5 * g)) 10).\nError: assert_equals: Color format is correct. expected "rgb(from rebeccapurple r calc(( * g) + ( * g)) )" but got "rgb(from rebeccapurple r calc( * g +  * g) )"
[FAIL] e.style['color'] = "rgb(from rebeccapurple r calc(b * .5 - g * .5) 10)" should set the property value
  Colors do not match.\nActual:   rgb(from rebeccapurple r calc(0.5 * b - 0.5 * g) 10)\nExpected: rgb(from rebeccapurple r calc((0.5 * b) - (0.5 * g)) 10).\nError: assert_equals: Color format is correct. expected "rgb(from rebeccapurple r calc(( * b) - ( * g)) )" but got "rgb(from rebeccapurple r calc( * b -  * g) )"
[FAIL] e.style['color'] = "rgba(from rebeccapurple r calc(g * .5 + g * .5) 10)" should set the property value
  Colors do not match.\nActual:   rgb(from rebeccapurple r calc(0.5 * g + 0.5 * g) 10)\nExpected: rgb(from rebeccapurple r calc((0.5 * g) + (0.5 * g)) 10).\nError: assert_equals: Color format is correct. expected "rgb(from rebeccapurple r calc(( * g) + ( * g)) )" but got "rgb(from rebeccapurple r calc( * g +  * g) )"
[FAIL] e.style['color'] = "rgba(from rebeccapurple r calc(b * .5 - g * .5) 10)" should set the property value
  Colors do not match.\nActual:   rgb(from rebeccapurple r calc(0.5 * b - 0.5 * g) 10)\nExpected: rgb(from rebeccapurple r calc((0.5 * b) - (0.5 * g)) 10).\nError: assert_equals: Color format is correct. expected "rgb(from rebeccapurple r calc(( * b) - ( * g)) )" but got "rgb(from rebeccapurple r calc( * b -  * g) )"
Harness: the test ran to completion.
